More fixes

+++ sdk/eventhubs/azure-messaging-eventhubs/test/eventhubs-stress-test/src/inc/argagg.hpp
@@ -1159,7 +1159,7 @@ inline parser_results parser::parse(int argc, const char** argv, bool posOnly) c
         size_t flag_len = arg_i_len;
         if (long_flag_arg != nullptr)
         {
-          flag_len = long_flag_arg - arg_i_cstr;
+          flag_len = static_cast<size_t>(long_flag_arg - arg_i_cstr);
         }
         std::string long_flag_str(arg_i_cstr, flag_len);
 
@@ -1422,7 +1422,7 @@ namespace convert {
     }
     else
     {
-      std::string arg_str(begin, s - begin);
+      std::string arg_str(begin, static_cast<size_t>(s - begin));
       out_arg = argagg::convert::arg<T>(arg_str.c_str());
       s += 1;
       return true;
